+++ app/src/main/java/com/example/app21try6/transaction/transactionselect/TransSelectModel.kt
@@ -1,6 +1,5 @@
 package com.example.app21try6.transaction.transactionselect
 
-import androidx.room.Ignore
 /*
 data  class TransSelectModel(
     var sub_product_id:Int =0,
@@ -14,7 +13,7 @@ data class TransSelectModel(
     var sub_product_id: Int = 0,
     var item_name: String = "",
     var item_price: Int = 0,
-    var trans_detail_id: Int = 0,
+    var trans_detail_id: Long = 0,
     var qty: Double = 0.0, // backing field
     var is_selected: Boolean = false
 )
